Add utilities module
- Add load, show and show_test to utilities function
