+++ main.py
@@ -1,21 +1,6 @@
 from modules.URL import URL
+from modules.utilities import load
 import sys
-
-
-def show(body):
-    in_tag = False
-    for c in body:
-        if c == "<":
-            in_tag = True
-        elif c == ">":
-            in_tag = False
-        elif not in_tag:
-            print(c, end="")
-
-
-def load(url):
-    body = url.request()
-    show(body)
 
 
 if __name__ == "__main__":
